+++ src/main/java/me/shepherd23333/projectex/ProjectEXCommon.java
@@ -3,6 +3,7 @@ package me.shepherd23333.projectex;
 import net.minecraft.entity.player.EntityPlayer;
 
 import javax.annotation.Nullable;
+import java.math.BigInteger;
 
 /**
  * @author LatvianModder
@@ -13,6 +14,6 @@ public class ProjectEXCommon {
         return null;
     }
 
-    public void updateEMC(long emc) {
+    public void updateEMC(BigInteger emc) {
     }
 }
